+++ doc/changes.rst
@@ -6,6 +6,7 @@ desisurvey change log
 ------------------
 
 * Fixes #18, #49, #54.
+* Improvements to surveymovie script.
 * Add progress columns to track fiber assignment and planning.
 * Add support for optional depth-first survey strategy.
 * Docs now auto-generated at http://desisurvey.readthedocs.io/en/latest/